requirements.txt
Django==1.8
Markdown==2.6.1
PyJWT==1.1.0
django-cors-headers==1.1.0
django-filter==0.9.2
djangorestframework==3.1.1
ipython==3.1.0
psycopg2==2.6
six==1.9.0
